@@ -26,9 +26,9 @@ function Footer() {
           </form>
         </div>
       </section>
-      <div class="footer-links">
+      <div className="footer-links">
         <div className="footer-link-wrapper">
-          <div class="footer-link-items">
+          <div className="footer-link-items">
             <h2>About Us</h2>
             <Link to="/sign-up">How it works</Link>
             <Link to="/">Testimonials</Link>
@@ -36,7 +36,7 @@ function Footer() {
             <Link to="/">Investors</Link>
             <Link to="/">Terms of Service</Link>
           </div>
-          <div class="footer-link-items">
+          <div className="footer-link-items">
             <h2>Contact Us</h2>
             <Link to="/">Contact</Link>
             <Link to="/">Support</Link>
@@ -45,14 +45,14 @@ function Footer() {
           </div>
         </div>
         <div className="footer-link-wrapper">
-          <div class="footer-link-items">
+          <div className="footer-link-items">
             <h2>Videos</h2>
             <Link to="/">Submit Video</Link>
             <Link to="/">Ambassadors</Link>
             <Link to="/">Agency</Link>
             <Link to="/">Influencer</Link>
           </div>
-          <div class="footer-link-items">
+          <div className="footer-link-items">
             <h2>Social Media</h2>
             <Link to="/">Instagram</Link>
             <Link to="/">Facebook</Link>
@@ -61,18 +61,18 @@ function Footer() {
           </div>
         </div>
       </div>
-      <section class="social-media">
-        <div class="social-media-wrap">
-          <div class="footer-logo">
+      <section className="social-media">
+        <div className="social-media-wrap">
+          <div className="footer-logo">
             <Link to="/" className="social-logo">
               CAMPERS NATION
               <FontAwesomeIcon icon="fa-solid fa-tree" />
             </Link>
           </div>
-          <small class="website-rights">Campers Nation © 2023</small>
-          <div class="social-icons">
+          <small className="website-rights">Campers Nation © 2023</small>
+          <div className="social-icons">
             <Link
-              class="social-icon-link facebook"
+              className="social-icon-link facebook"
               to="/"
               target="_blank"
               aria-label="Facebook"
@@ -80,7 +80,7 @@ function Footer() {
               <FontAwesomeIcon icon="fa-brands fa-facebook" />
             </Link>
             <Link
-              class="social-icon-link instagram"
+              className="social-icon-link instagram"
               to="/"
               target="_blank"
               aria-label="Instagram"
@@ -88,7 +88,7 @@ function Footer() {
               <FontAwesomeIcon icon="fa-brands fa-instagram" />
             </Link>
             <Link
-              class="social-icon-link youtube"
+              className="social-icon-link youtube"
               to="/"
               target="_blank"
               aria-label="Youtube"
@@ -96,7 +96,7 @@ function Footer() {
               <FontAwesomeIcon icon="fa-brands fa-youtube" />
             </Link>
             <Link
-              class="social-icon-link twitter"
+              className="social-icon-link twitter"
               to="/"
               target="_blank"
               aria-label="Twitter"
@@ -104,7 +104,7 @@ function Footer() {
               <FontAwesomeIcon icon="fa-brands fa-twitter" />
             </Link>
             <Link
-              class="social-icon-link twitter"
+              className="social-icon-link twitter"
               to="/"
               target="_blank"
               aria-label="LinkedIn"
